refactor RestClient to RestClientSingleton
